+++ package.json
@@ -2,9 +2,9 @@
   "name": "fetch-rewards-coding-exercise",
   "version": "1.0.0",
   "description": "Fetch Rewards Coding Exercise",
-  "main": "source/server.ts",
+  "main": "src/server.ts",
   "scripts": {
-    "start": "nodemon src/server.ts",
+    "dev": "nodemon src/server.ts",
     "test": "mocha -r ts-node/register src/**/*.spec.ts",
     "build": "rm -rf build && prettier --write source/ && tsc"
   },
