+++ src/controllers/main.ts
@@ -3,14 +3,22 @@ import { Bill, Transaction } from "../interface/main";
 import helper from "../utils/helper";
 import server from "../server";
 import config from "../config/config";
+
+//  set the two global variables
 const transactions: Transaction[] = [];
 const balance = {};
 
 const addTransactions = (req: Request, res: Response) => {
   try {
     const { payer, points, timestamp } = req.body;
-    const transaction: Transaction = { payer, points, timestamp };
-    helper.addTransactionHelper(transactions, transaction);
+    // validate the parameters
+    let valid_timestamp = new Date(timestamp).toISOString();
+    if (typeof payer !== "string" || typeof points !== "number") {
+      throw "invalid parameters";
+    }
+    // create & add a new transaction
+    const transaction: Transaction = { payer, points, timestamp: valid_timestamp };
+    helper.addTransactionHelper(transactions, transaction, balance);
     return res.status(200).json({
       message: "added a transaction successfully!",
     });
@@ -24,6 +32,11 @@ const addTransactions = (req: Request, res: Response) => {
 const spendPoints = async (req: Request, res: Response) => {
   try {
     const { points } = req.body;
+    // validate the parameters
+    if (typeof points !== "number") {
+      throw "invalid parameters";
+    }
+    // generate bills & modify the transactions if necessary
     const bills: Bill[] = helper.spendPointHelper(transactions, points, balance);
     return res.status(200).json(bills);
   } catch (err) {
@@ -35,7 +48,6 @@ const spendPoints = async (req: Request, res: Response) => {
 
 const getBalence = async (req: Request, res: Response) => {
   try {
-    helper.getBalenceHelper(transactions, balance);
     return res.status(200).json(balance);
   } catch (err) {
     return res.status(404).json({
