+++ src/routes/main.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import controller from "../controllers/controller";
+import controller from "../controllers/main";
 
 const router = express.Router();
