+++ src/utils/helper.ts
@@ -1,22 +1,48 @@
 import { Transaction, Bill } from "../interface/main";
 
-const addTransactionHelper = (transactions: Transaction[], transaction: Transaction) => {
+const addTransactionHelper = (transactions: Transaction[], transaction: Transaction, balance: any) => {
   const N = transactions.length;
+  balance[transaction.payer] = (balance[transaction.payer] || 0) + transaction.points;
   if (N === 0) {
-    // console.log("bill132143");
     transactions.push(transaction);
-    // console.log("bill132143", transactions);
     return;
   }
   for (let i = 0; i < N; i++) {
     if (transactions[i].timestamp > transaction.timestamp) {
       transactions.splice(i, 0, transaction);
-
       return;
     }
   }
   transactions.splice(N - 1, 0, transaction);
-  // console.log("bill", transactions);
+  console.log(balance);
+};
+
+const spendPointHelper = (transactions: Transaction[], pointAmount: number, balance: any) => {
+  let bills: Bill[] = [];
+  while (transactions.length) {
+    if (pointAmount > 0) {
+      const transaction: Transaction | undefined = transactions.shift();
+
+      if (transaction) {
+        let bPints: number = 0;
+        bPints = Math.min(pointAmount, transaction.points);
+        let bill: Bill = { payer: transaction.payer, points: bPints };
+        // To add the bill to bills list for response
+        addBill(bills, bill);
+        // To update the balance
+        balance[transaction.payer] -= bPints;
+        if (pointAmount < transaction.points) {
+          transaction.points -= bPints;
+          transactions.unshift(transaction);
+        }
+
+        pointAmount -= bPints;
+      }
+    } else {
+      break;
+    }
+  }
+  return bills;
 };
 
 const addBill = (bills: Bill[], bill: Bill) => {
@@ -32,49 +58,4 @@ const addBill = (bills: Bill[], bill: Bill) => {
   }
 };
 
-const spendPointHelper = (transactions: Transaction[], pointAmount: number, balance: any) => {
-  let bills: Bill[] = [];
-
-  while (transactions.length) {
-    if (pointAmount > 0) {
-      const transaction: Transaction | undefined = transactions.shift();
-
-      if (transaction) {
-        if (!(transaction.payer in balance)) {
-          balance[transaction.payer] = 0;
-        }
-        if (balance[transaction.payer] != 0) {
-          balance[transaction.payer] = 0;
-        }
-        let bPints: number = 0;
-        bPints = Math.min(pointAmount, transaction.points);
-        let bill: Bill = { payer: transaction.payer, points: bPints };
-        // console.log(bill);
-        addBill(bills, bill);
-        if (pointAmount < transaction.points) {
-          transaction.points -= bPints;
-          transactions.unshift(transaction);
-        }
-        pointAmount -= bPints;
-      }
-    } else {
-      break;
-    }
-  }
-
-  return bills;
-};
-
-const getBalenceHelper = (transactions: Transaction[], balance: any) => {
-  // let balance: any = {};
-  transactions.forEach((trans) => {
-    if (trans.payer in balance) {
-      balance[trans.payer] += trans.points;
-    } else {
-      balance[trans.payer] = trans.points;
-    }
-  });
-  return balance;
-};
-
-export default { addTransactionHelper, spendPointHelper, getBalenceHelper };
+export default { addTransactionHelper, spendPointHelper };
